refactor(core): Move `ApplicationError` to `@n8n/errors` (#17667)
