+++ packages/@n8n/errors/src/application.error.ts
@@ -1,7 +1,7 @@
 import type { Event } from '@sentry/node';
 import callsites from 'callsites';
 
-import type { ErrorLevel, ReportingOptions } from './error.types';
+import type { ErrorLevel, ReportingOptions } from './types';
 
 /**
  * @deprecated Use `UserError`, `OperationalError` or `UnexpectedError` instead.
@@ -26,9 +26,11 @@ export class ApplicationError extends Error {
 
 		try {
 			const filePath = callsites()[2].getFileName() ?? '';
+			// eslint-disable-next-line no-useless-escape
 			const match = /packages\/([^\/]+)\//.exec(filePath)?.[1];
 
 			if (match) this.tags.packageName = match;
+			// eslint-disable-next-line no-empty
 		} catch {}
 	}
 }
