+++ packages/@n8n/errors/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { ApplicationError } from './application.error';
+export * from './types';
