+++ packages/@n8n/nodes-langchain/nodes/trigger/ChatTrigger/error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class ChatTriggerAuthorizationError extends ApplicationError {
 	constructor(
