+++ packages/@n8n/task-runner/src/js-task-runner/errors/task-cancelled-error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class TaskCancelledError extends ApplicationError {
 	constructor(reason: string) {
