+++ packages/@n8n/task-runner/src/js-task-runner/errors/timeout-error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class TimeoutError extends ApplicationError {
 	description: string;
