+++ packages/@n8n/task-runner/src/js-task-runner/errors/unsupported-function.error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 /**
  * Error that indicates that a specific function is not available in the
