+++ packages/core/src/errors/abstract/binary-data.error.ts
@@ -1,3 +1,3 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export abstract class BinaryDataError extends ApplicationError {}
