+++ packages/core/src/errors/abstract/filesystem.error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export abstract class FileSystemError extends ApplicationError {
 	constructor(message: string, filePath: string) {
