+++ packages/core/src/errors/invalid-execution-metadata.error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class InvalidExecutionMetadataError extends ApplicationError {
 	constructor(
