+++ packages/core/src/errors/invalid-mode.error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 import { CONFIG_MODES } from '../binary-data/utils';
