+++ packages/core/src/execution-engine/node-execution-context/hook-context.ts
@@ -1,3 +1,4 @@
+import { ApplicationError } from '@n8n/errors';
 import type {
 	ICredentialDataDecryptedObject,
 	INode,
@@ -9,7 +10,6 @@ import type {
 	IWebhookData,
 	WebhookType,
 } from 'n8n-workflow';
-import { ApplicationError } from 'n8n-workflow';
 
 import { NodeExecutionContext } from './node-execution-context';
 import { getRequestHelperFunctions } from './utils/request-helper-functions';
