+++ packages/core/src/execution-engine/node-execution-context/utils/__tests__/normalize-items.test.ts
@@ -1,5 +1,5 @@
+import { ApplicationError } from '@n8n/errors';
 import type { IBinaryData, INodeExecutionData } from 'n8n-workflow';
-import { ApplicationError } from 'n8n-workflow';
 
 import { normalizeItems } from '../normalize-items';
