+++ packages/nodes-base/nodes/Code/ValidationError.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class ValidationError extends ApplicationError {
 	description = '';
