+++ packages/nodes-base/nodes/Code/errors/WrappedExecutionError.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export type WrappableError = Record<string, unknown>;
