+++ packages/nodes-base/nodes/Code/test/Code.node.test.ts
@@ -3,7 +3,7 @@ import { NodeTestHarness } from '@nodes-testing/node-test-harness';
 import { anyNumber, mock } from 'jest-mock-extended';
 import { normalizeItems } from 'n8n-core';
 import type { IExecuteFunctions, IWorkflowDataProxyData } from 'n8n-workflow';
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 import { Code } from '../Code.node';
 import { ValidationError } from '../ValidationError';
