+++ packages/nodes-base/nodes/Merge/v2/utils.ts
@@ -10,7 +10,7 @@ import type {
 	INodeExecutionData,
 	IPairedItemData,
 } from 'n8n-workflow';
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 import { fuzzyCompare, preparePairedItemDataArray } from '@utils/utilities';
