+++ packages/nodes-base/nodes/Venafi/Datacenter/GenericFunctions.ts
@@ -1,5 +1,5 @@
 import get from 'lodash/get';
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 import type {
 	IDataObject,
 	IExecuteFunctions,
