+++ packages/nodes-base/nodes/Webhook/error.ts
@@ -1,4 +1,4 @@
-import { ApplicationError } from 'n8n-workflow';
+import { ApplicationError } from '@n8n/errors';
 
 export class WebhookAuthorizationError extends ApplicationError {
 	constructor(
